+++ libs/spark/CHANGELOG.md
@@ -32,6 +32,12 @@ _This section details previews of breaking changes or experimental features that
   - Initial implementation.
 - **Unstable_Paper**
   - Initial implementation.
+- **Unstable_Select**
+  - Props API changes:
+    - `preventMultipleOverflow`: added; values `'large' | 'medium'` where `'large'` is default.
+  - Styles: match new menu (paper) specifications.
+- **Unstable_TextField**
+  - See **Unstable_Select**
 
 ## [v1.0.0-alpha.10](https://github.com/prenda-school/prenda-spark/compare/v1.0.0-alpha.9...v1.0.0-alpha.10) (2022-06-22)